+++ src/main/java/com/gladkaya/shape/parser/DataParser.java
@@ -10,9 +10,9 @@ public class DataParser {
     private final static Logger logger = LogManager.getLogger();
     private final static String STRING_REGEXP = "\\s";
 
-    public float[] parseTriangle(String correctLine) {
+    public double[] parseTriangle(String correctLine) {
         String[] stringTriangle = correctLine.split(STRING_REGEXP);
-        float[] triangleLine = new float[stringTriangle.length];
+        double[] triangleLine = new double[stringTriangle.length];
         for (int i = 0; i < stringTriangle.length; i++) {
             triangleLine[i] = Float.parseFloat(stringTriangle[i]);
         }
